requirements.txt
fastapi
uvicorn
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings
sqlalchemy
psycopg2-binary
qdrant-client
langchain
langchain-openai
langchain-community
openai
tiktoken
ragas
pytest
httpx
python-dotenv
alembic
arize-phoenix
langsmith
pypdf
langchain-google-vertexai
